# backend/swagger-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "根据ID获取用户信息",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "更新用户信息",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在"
          },
          "409": {
            "description": "邮箱已被其他用户使用"
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "删除用户",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/orders/user/{userId}": {
      "get": {
        "operationId": "OrdersController_getUserOrders",
        "summary": "获取用户订单列表",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "订单"
        ]
      }
    },
...
    "/admin/auth/login": {
      "post": {
        "operationId": "AdminAuthController_login",
        "summary": "管理员登录",
        "description": "管理员用户登录系统",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功"
          },
          "401": {
            "description": "账号或密码错误"
          }
        },
        "tags": [
          "后台管理-认证"
        ]
      }
    },
    "/admin/auth/profile": {
      "get": {
        "operationId": "AdminAuthController_getProfile",
        "summary": "获取管理员信息",
        "description": "获取当前登录管理员的基本信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-认证"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/auth/change-password": {
      "post": {
        "operationId": "AdminAuthController_changePassword",
        "summary": "修改密码",
        "description": "修改当前登录管理员的密码",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码修改成功"
          },
          "400": {
            "description": "原密码错误"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-认证"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/auth/validate-token": {
      "get": {
        "operationId": "AdminAuthController_validateToken",
        "summary": "验证令牌",
        "description": "验证JWT令牌是否有效",
        "parameters": [],
        "responses": {
          "200": {
            "description": "令牌有效"
          },
          "401": {
            "description": "令牌无效"
          }
        },
        "tags": [
          "后台管理-认证"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users": {
      "post": {
        "operationId": "AdminUserController_create",
        "summary": "创建管理员用户",
        "description": "创建新的管理员用户账号",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户创建成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "角色不存在"
          },
          "409": {
            "description": "用户名已存在"
          }
        },
        "tags": [
          "后台管理-用户管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "AdminUserController_findAll",
        "summary": "获取用户列表",
        "description": "获取管理员用户列表，支持分页和搜索",
        "parameters": [
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "description": "角色ID筛选",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态筛选：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "搜索关键词（用户名、真实姓名）",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量，默认10",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，默认1",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-用户管理"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminUserController_findOne",
        "summary": "获取用户详情",
        "description": "根据ID获取管理员用户的详细信息",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "tags": [
          "后台管理-用户管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "AdminUserController_update",
        "summary": "更新用户信息",
        "description": "更新管理员用户的基本信息",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          },
          "409": {
            "description": "用户名已存在"
          }
        },
        "tags": [
          "后台管理-用户管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminUserController_remove",
        "summary": "删除用户",
        "description": "删除指定的管理员用户账号",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "400": {
            "description": "不能删除自己"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "tags": [
          "后台管理-用户管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{id}/reset-password": {
      "post": {
        "operationId": "AdminUserController_resetPassword",
        "summary": "重置密码",
        "description": "重置指定用户的密码为默认密码",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "密码重置成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "tags": [
          "后台管理-用户管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{id}/status": {
      "patch": {
        "operationId": "AdminUserController_toggleStatus",
        "summary": "启用/禁用用户",
        "description": "启用或禁用指定的管理员用户账号",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "description": "状态：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "400": {
            "description": "不能禁用自己"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "tags": [
          "后台管理-用户管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "二手交易系统 API",
    "description": "二手交易系统后端接口文档，包含前台用户接口和后台管理接口",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "address": {
            "type": "string",
            "description": "地址"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "created_at",
          "updated_at"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "管理员账号",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "admin123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "address": {
            "type": "string",
            "description": "地址"
          }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {}
      },
...
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "原密码",
            "example": "oldPassword123"
          },
          "newPassword": {
            "type": "string",
            "description": "新密码",
            "example": "newPassword456"
          },
          "confirmPassword": {
            "type": "string",
            "description": "确认新密码",
            "example": "newPassword456"
          }
        },
        "required": [
          "oldPassword",
          "newPassword",
          "confirmPassword"
        ]
      },
      "CreateAdminUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "管理员账号",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "password123"
          },
          "realName": {
            "type": "string",
            "description": "真实姓名",
            "example": "张三"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "roleId": {
            "type": "number",
            "description": "角色ID",
            "example": 1
          }
        },
        "required": [
          "username",
          "password",
          "roleId"
        ]
      },
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "状态：1启用 0禁用",
            "example": 1
          },
          "realName": {
            "type": "string",
            "description": "真实姓名",
            "example": "李四"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138001"
          },
          "roleId": {
            "type": "number",
            "description": "角色ID",
            "example": 2
          }
        }
      }
    }
  }
}